+++ magpie/db.py
@@ -1,10 +1,9 @@
 #!/usr/bin/python
 # -*- coding: utf-8 -*-
 from magpie import constants
+from magpie.common import get_settings_from_config_ini
 from magpie.definitions.alembic_definitions import *
 from magpie.definitions.sqlalchemy_definitions import *
-# noinspection PyCompatibility
-import configparser
 import transaction
 import inspect
 import zope.sqlalchemy
@@ -78,14 +77,6 @@ def get_db_session_from_settings(settings):
 def get_db_session_from_config_ini(config_ini_path, ini_main_section_name='app:magpie_app'):
     settings = get_settings_from_config_ini(config_ini_path, ini_main_section_name)
     return get_db_session_from_settings(settings)
-
-
-def get_settings_from_config_ini(config_ini_path, ini_main_section_name='app:magpie_app'):
-    parser = configparser.ConfigParser()
-    parser.optionxform = lambda option: option  # preserve case of config (ziggurat requires it for 'User' model)
-    parser.read([config_ini_path])
-    settings = dict(parser.items(ini_main_section_name))
-    return settings
 
 
 def run_database_migration():