+++ HISTORY.rst
@@ -10,6 +10,7 @@ History
 * fix many invalid or erroneous swagger specifications
 * attempt db creation on first migration if not existing
 * add continuous integration testing and deployment
+* reduce excessive sqlalchemy logging using `MAGPIE_LOG_LEVEL >= INFO`
 
 0.8.x
 ---------------------